library: fix exception when parsing a recovered .plt
In commit 440aface1540 ("library: add support for PLT calls
with Intel CET"), we added detection logic for .plt.sec
sections but the change would only correctly work on a Section
object as the newly added code accesses a .name member of
the plt object.

Fix this by checking if we actually have a Section first.

Reported-by: Stefan Eschenbacher <stefan.eschenbacher@fau.de>