store: sort uprobe strings by library and address
This way, we get the same order every time we analyze the same
binaries which makes them more easy to compare and to take uprobe
measurements across different JSON files.

Additionally, remove to 'u:' prefix for which a kernel patch is
required in favor of the natively supported 'p:' prefix.