library: allow extraction of one name per used/unused function only
Functions can be aliased to the same address so we could return a
greater number of used or unused functions than there are distinct
functions in the binary. This could be problematic if we compare
data between different approaches, so allow restricting the
collected names to one name per function only.