library, store, calls: partly handle dlsym
If we detect a call to a function containing 'dlsym' in its name,
try to extract the target string from the underlying file and
store this function name as referenced from the currently
disassembled function. This way, the transitive call selection
can later pick this edge up and mark the target functin as used.

Additionally, extract the target strings for 'dlopen'-named
functions in order to allow the identification of dynamically
loaded libraries.

This currently only works on x86_64.