interface_calls: shuffle checks for references around
When we're checking for references to known functions or
objects, we actually need to check for objects first as any
relocation pointing to the beginning of an object but
referencing a function (e.g., a function pointer as the
first member of a struct or class) might otherwise be
identified as a reference to a single function only.

Especially in the case that ALL_FUNCTIONS_FROM_OBJECTS is
set we definitely want the reference to an object to include
all functions and not just the first. Note that this can
lead to a slight overapproximation in case the original code
really only referenced the first element but from experience
we know that such references usually mean the object.

Additionally, include references to members of exports_plt and
the more recent reloc_to_{local,exported} as we didn't account
for those in the reference case previously.