library: also record _RELATIVE relocations pointing directly at functions
As R_X86_64_RELATIVE relocations can also directly point at
functions, we need to check for such accesses as well. This makes
the detection of function references a lot better for librsvg, for
example, as many references to functions in this Rust library seem
to use this pattern (e.g., references to __rust_alloc).