library: register additional offset for vtable objects
As C++ vtables generated by g++ start with a qword 0 (on x86_64)
and a qword for the RTTI for the object, the compiler knows that
it can also access <vtable_offset>+16 as the base address for
virtual function pointers. In order to detect such references
from the disassembly, note these offsets as local objects as
well and have them point at the surrounding vtable symbol table
object through Library.object_to_objects.