library, interface_calls: better treatment for local objects
Objects do not need to be referenced directly by their base address
when the compiler knows the actual offset of the member the code
is accessing. To cover this case, use bisection (as in the 'call
into a function' implementation) to check if a data access is
inside the range of a known object.

Additionally, filter out some stray references when parsing the
symbol tables.