library: extract constant escaping and pattern construction from loop
The pattern required for the lookup is constant over all iterations
of the loop, so pull it out to avoid constructing it over and over
again. This saves roughly 1.5 minutes for looking up 6000 traced
probes during the analysis of a MariaDB system with containing 24000
local funcions across all ELF files.