library: add COPIED_BASE environment variable
This allows to redirect the (usually absolute) path in a PT_INTERP
program header into a copied file system tree where all binaries
as well as the loader are located. This allows to distinguish
between a specific loader to the application and the system-wide
loader on the analysis host system.