library: reduce calls to ELFFile.address_offsets()
Instead of calling ELFFile.address_offsets() for every relocation,
save the range of the DT_INIT_ARRAY and DT_FINI_ARRAY as virtual
addresses and check the relocation entries directly against this
range. For executables and shared objects, the r_offset value in
the relocation entry should always be a virtual address.

While at it, unify the implementations of checking the relocations
against the INIT_ARRAY and FINI_ARRAY ranges inside
Library.parse_rela_dyn() and adapt parsing the arrays in
Library._process_function_array().